Migrate babel plugin tests to use ava
