+++ packages/babel/package.json
@@ -9,7 +9,7 @@
   "bugs": {
     "url": "https://github.com/rollup/rollup-plugin-babel/issues"
   },
-  "main": "dist/rollup-plugin-babel.cjs.js",
+  "main": "./dist/rollup-plugin-babel.cjs.js",
   "engines": {
     "node": ">=10"
   },
@@ -19,7 +19,7 @@
     "prepare": "npm run build",
     "pretest": "npm run build",
     "preversion": "npm test",
-    "test": "mocha"
+    "test": "ava"
   },
   "files": [
     "dist",
@@ -46,16 +46,23 @@
     "@babel/plugin-syntax-dynamic-import": "^7.7.4",
     "@babel/plugin-transform-runtime": "^7.7.4",
     "@babel/preset-env": "^7.7.4",
-    "mocha": "^6.2.2",
+    "ava": "^2.4.0",
     "rimraf": "^2.6.3",
     "rollup": "^1.27.5",
     "rollup-plugin-json": "^4.0.0",
-    "source-map": "^0.6.1",
-    "source-map-support": "^0.5.16"
+    "source-map": "^0.6.1"
+  },
+  "ava": {
+    "files": [
+      "!**/fixtures/**",
+      "!**/helpers/**",
+      "!**/recipes/**",
+      "!**/types.ts"
+    ]
   },
   "contributors": [
     "Bogdan Chadkin <trysound@yandex.ru>",
     "Mateusz Burzyński <mateuszburzynski@gmail.com> (https://github.com/Andarist)"
   ],
-  "module": "dist/rollup-plugin-babel.esm.js"
+  "module": "./dist/rollup-plugin-babel.esm.js"
 }
