+++ packages/babel/test/fixtures/.babelrc
