+++ packages/babel/test/fixtures/basic/main.js
