+++ packages/babel/test/fixtures/checks/foo/.babelrc
