+++ packages/babel/test/fixtures/checks/foo/index.js
