+++ packages/babel/test/fixtures/checks/main.js
