+++ packages/babel/test/fixtures/class-loose/.babelrc
