+++ packages/babel/test/fixtures/class-loose/main.js
