+++ packages/babel/test/fixtures/class/main.js
