+++ packages/babel/test/fixtures/config-file/config.json
