+++ packages/babel/test/fixtures/config-file/main.js
