+++ packages/babel/test/fixtures/exclusions/foo.js
