+++ packages/babel/test/fixtures/extensions-custom/ext.es6
