+++ packages/babel/test/fixtures/extensions-custom/ext.mjs
