+++ packages/babel/test/fixtures/extensions-custom/main.js
