+++ packages/babel/test/fixtures/extensions-default/ext.es6
