+++ packages/babel/test/fixtures/extensions-default/ext.json
