+++ packages/babel/test/fixtures/extensions-default/ext.jsx
