+++ packages/babel/test/fixtures/extensions-default/main.js
