+++ packages/babel/test/fixtures/external-helpers/bar.js
