+++ packages/babel/test/fixtures/external-helpers/foo.js
