+++ packages/babel/test/fixtures/ignored-file/.babelrc
