+++ packages/babel/test/fixtures/ignored-file/ignored.js
