+++ packages/babel/test/fixtures/ignored-file/main.js
