+++ packages/babel/test/fixtures/named-function-helper/_config.js
