+++ packages/babel/test/fixtures/named-function-helper/main.js
