+++ packages/babel/test/fixtures/no-class-transformer/.babelrc
