+++ packages/babel/test/fixtures/no-class-transformer/main.js
