+++ packages/babel/test/fixtures/preserve-modules/dep.js
