+++ packages/babel/test/fixtures/preserve-modules/main.js
