+++ packages/babel/test/fixtures/proposal-decorators/.babelrc
