+++ packages/babel/test/fixtures/proposal-decorators/main.js
