+++ packages/babel/test/fixtures/runtime-helpers-commonjs/.babelrc
