+++ packages/babel/test/fixtures/runtime-helpers-commonjs/main.js
