+++ packages/babel/test/fixtures/typeof/bar.js
