+++ packages/babel/test/fixtures/typeof/foo.js
