+++ packages/babel/test/fixtures/typeof/main.js
